Backend/requirements.txt
aniso8601==9.0.1
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
configparser==5.0.2
cryptography==38.0.4
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==3.4
itsdangerous==2.1.2
joblib==1.2.0
mysql-connector-python==8.0.32
numpy==1.24.2
pandas==2.0.0
pip==22.3.1
protobuf==3.20.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.1
schedule==1.1.0
scikit-learn==1.2.2
scipy==1.10.1
setuptools==65.6.3
six==1.16.0
sklearn==0.0.post1
SQLAlchemy==1.3.18
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==1.26.14
Werkzeug==2.2.3
wheel==0.38.4
